+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to
 - ⚡️(frontend) limit unnecessary re-renders #1510
 - 📝(legal) update terms of service
 - 💄(frontend) render Avatar initials in uppercase
+- 💄(frontend) improve participant name rendering in the list
 
 ## Fixed
 
@@ -34,7 +35,6 @@ and this project adheres to
 - 🐛(backend) preserve recording metadata when updating room access
 - 🐛(backend) allow any string as sub in the API serializer
 - 🐛(frontend) fall back to user.full_name on request-entry
-
 
 ## [1.24.0] - 2026-07-21